refactor useComputationNotification
- dont store a computed state from RunningStatus
- remove useComputationNotification
- remove useEffect in runButtonContainer used to syncronize notification state
