+++ src/hooks/use-computation-notification-count.ts
@@ -6,31 +6,75 @@
  */
 import { useSelector } from 'react-redux';
 import { ReduxState } from '../redux/reducer.type';
+import ComputingType from 'components/computing-status/computing-type';
+import RunningStatus from 'components/utils/running-status';
+import { isNodeBuilt } from 'components/graph/util/model-functions';
 
 export const useComputationNotificationCount = () => {
-    const loadflowNotif = useSelector(
-        (state: ReduxState) => state.loadflowNotif
+    const currentNode = useSelector(
+        (state: ReduxState) => state.currentTreeNode
     );
 
-    const saNotif = useSelector((state: ReduxState) => state.saNotif);
-
-    const sensiNotif = useSelector((state: ReduxState) => state.sensiNotif);
-
-    const allBusesshortCircuitNotif = useSelector(
-        (state: ReduxState) => state.allBusesShortCircuitNotif
+    const loadFlowStatus = useSelector(
+        (state: ReduxState) => state.computingStatus[ComputingType.LOADFLOW]
     );
 
-    const oneBusShortCircuitNotif = useSelector(
-        (state: ReduxState) => state.oneBusShortCircuitNotif
+    const securityAnalysisStatus = useSelector(
+        (state: ReduxState) =>
+            state.computingStatus[ComputingType.SECURITY_ANALYSIS]
     );
 
-    const voltageInitNotif = useSelector(
-        (state: ReduxState) => state.voltageInitNotif
+    const sensitivityAnalysisStatus = useSelector(
+        (state: ReduxState) =>
+            state.computingStatus[ComputingType.SENSITIVITY_ANALYSIS]
     );
 
-    const dynamicSimulationNotif = useSelector(
-        (state: ReduxState) => state.dynamicSimulationNotif
+    const oneBusallBusesShortCircuitStatus = useSelector(
+        (state: ReduxState) =>
+            state.computingStatus[ComputingType.ONE_BUS_SHORTCIRCUIT_ANALYSIS]
     );
+
+    const allBusesShortCircuitStatus = useSelector(
+        (state: ReduxState) =>
+            state.computingStatus[ComputingType.ALL_BUSES_SHORTCIRCUIT_ANALYSIS]
+    );
+
+    const dynamicSimulationStatus = useSelector(
+        (state: ReduxState) =>
+            state.computingStatus[ComputingType.DYNAMIC_SIMULATION]
+    );
+
+    const voltageInitStatus = useSelector(
+        (state: ReduxState) => state.computingStatus[ComputingType.VOLTAGE_INIT]
+    );
+
+    const loadflowNotif =
+        isNodeBuilt(currentNode) &&
+        (loadFlowStatus === RunningStatus.SUCCEED ||
+            loadFlowStatus === RunningStatus.FAILED);
+    const saNotif =
+        isNodeBuilt(currentNode) &&
+        (securityAnalysisStatus === RunningStatus.SUCCEED ||
+            securityAnalysisStatus === RunningStatus.FAILED);
+    const sensiNotif =
+        isNodeBuilt(currentNode) &&
+        sensitivityAnalysisStatus === RunningStatus.SUCCEED;
+    const allBusesshortCircuitNotif =
+        isNodeBuilt(currentNode) &&
+        allBusesShortCircuitStatus === RunningStatus.SUCCEED;
+
+    const oneBusShortCircuitNotif =
+        isNodeBuilt(currentNode) &&
+        oneBusallBusesShortCircuitStatus === RunningStatus.SUCCEED;
+    const voltageInitNotif =
+        (isNodeBuilt(currentNode) &&
+            voltageInitStatus === RunningStatus.SUCCEED) ||
+        voltageInitStatus === RunningStatus.FAILED;
+
+    const dynamicSimulationNotif =
+        (isNodeBuilt(currentNode) &&
+            dynamicSimulationStatus === RunningStatus.SUCCEED) ||
+        dynamicSimulationStatus === RunningStatus.FAILED;
 
     return [
         loadflowNotif,
