+++ src/redux/reducer.js
@@ -218,13 +218,6 @@ const initialState = {
     showAuthenticationRouterLogin: false,
     studyUpdated: { force: 0, eventData: {} },
     mapDataLoading: false,
-    loadflowNotif: false,
-    saNotif: false,
-    voltageInitNotif: false,
-    sensiNotif: false,
-    allBusesShortCircuitNotif: false,
-    oneBusShortCircuitNotif: false,
-    dynamicSimulationNotif: false,
     fullScreenDiagram: null,
     allDisplayedColumnsNames: TABLES_COLUMNS_NAMES_JSON,
     allLockedColumnsNames: [],
